fix project setup
