+++ LearningPlatform/LearningPlatform.Application/Interfaces/ICourseRepository.cs
@@ -1,0 +1,11 @@
+﻿using LearningPlatform.Core.Models;
+
+namespace LearningPlatform.Application.Interfaces;
+public interface ICourseRepository
+{
+    Task Create(Course course);
+    Task Delete(Guid id);
+    Task<List<Course>> Get();
+    Task<Course> GetById(Guid id);
+    Task Update(Guid id, string title, string description, decimal price);
+}
