+++ LearningPlatform/LearningPlatform.Application/Interfaces/ILessonsRepository.cs
@@ -1,0 +1,11 @@
+﻿using LearningPlatform.Core.Models;
+
+namespace LearningPlatform.Application.Interfaces;
+public interface ILessonsRepository
+{
+    Task Create(Lesson lesson);
+    Task Delete(Guid id);
+    Task<List<Lesson>> Get(Guid courseId);
+    Task<Lesson> GetById(Guid id);
+    Task Update(Guid id, string title, string description, string videoLink, string lessonText);
+}
