+++ LearningPlatform/LearningPlatform.Application/Interfaces/IUsersRepository.cs
@@ -1,0 +1,8 @@
+﻿using LearningPlatform.Core.Models;
+
+namespace LearningPlatform.Application.Interfaces;
+public interface IUsersRepository
+{
+    Task Create(User user);
+    Task<User> GetByEmail(string email);
+}
