+++ LearningPlatform/LearningPlatform.Application/Services/CoursesService.cs
@@ -1,0 +1,43 @@
+﻿using LearningPlatform.Application.Interfaces;
+using LearningPlatform.Core.Models;
+
+namespace LearningPlatform.Application.Services;
+
+public class CoursesService
+{
+    private readonly ICourseRepository _courseRepository;
+
+    public CoursesService(ICourseRepository courseRepository)
+    {
+        _courseRepository = courseRepository;
+    }
+
+    public async Task CreateCourse(Course course)
+    {
+        await _courseRepository.Create(course);
+    }
+
+    public async Task<List<Course>> GetCourses()
+    {
+        return await _courseRepository.Get();
+    }
+
+    public async Task<Course> GetCourseById(Guid id)
+    {
+        return await _courseRepository.GetById(id);
+    }
+
+    public async Task UpdateCourse(
+        Guid id,
+        string title,
+        string description,
+        decimal price)
+    {
+        await _courseRepository.Update(id, title, description, price);
+    }
+
+    public async Task DeleteCourse(Guid id)
+    {
+        await _courseRepository.Delete(id);
+    }
+}
