+++ LearningPlatform/LearningPlatform.Application/Services/PasswordHasher.cs
@@ -1,0 +1,9 @@
+﻿namespace LearningPlatform.Application.Services;
+public static class PasswordHasher
+{
+    public static string Generate(string password)
+        => BCrypt.Net.BCrypt.EnhancedHashPassword(password);
+
+    public static bool Verify(string password, string hashedPassword)
+        => BCrypt.Net.BCrypt.EnhancedVerify(password, hashedPassword);
+}
