+++ LearningPlatform/LearningPlatform.Persistance/Configurations/CourseConfiguration.cs
