+++ LearningPlatform/LearningPlatform.Persistance/Configurations/LessonConfiguration.cs
