+++ LearningPlatform/LearningPlatform.Persistance/Entities/CourseEntity.cs
