+++ LearningPlatform/LearningPlatform.Persistance/Entities/UserEntity.cs
