+++ LearningPlatform/LearningPlatform.sln
@@ -7,14 +7,14 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "LearningPlatform.API", "Lea
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "LearningPlatform.Core", "LearningPlatform.Core\LearningPlatform.Core.csproj", "{7DD7F3D0-80EC-4EE0-8384-FB7C39BCCDDC}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "LearningPlatform.Persistance", "LearningPlatform.DataAccess.Postgres\LearningPlatform.Persistance.csproj", "{260BBE02-292D-4897-B20D-869B452A4C0A}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "LearningPlatforn.Application", "LearningPlatforn.Application\LearningPlatforn.Application.csproj", "{773193DB-3981-4577-9108-4D4147C873EB}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{6C368082-F625-4871-B6AB-E8767ABE1F8A}"
 	ProjectSection(SolutionItems) = preProject
 		.editorconfig = .editorconfig
 	EndProjectSection
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "LearningPlatform.Persistance", "LearningPlatform.Persistance\LearningPlatform.Persistance.csproj", "{8F36B825-22A5-4301-AF04-B56D4E2DB4EF}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "LearningPlatform.Application", "LearningPlatform.Application\LearningPlatform.Application.csproj", "{AF86C3F1-5248-4ED1-98A5-BBDC95A55560}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -30,14 +30,14 @@ Global
 		{7DD7F3D0-80EC-4EE0-8384-FB7C39BCCDDC}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{7DD7F3D0-80EC-4EE0-8384-FB7C39BCCDDC}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{7DD7F3D0-80EC-4EE0-8384-FB7C39BCCDDC}.Release|Any CPU.Build.0 = Release|Any CPU
-		{260BBE02-292D-4897-B20D-869B452A4C0A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{260BBE02-292D-4897-B20D-869B452A4C0A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{260BBE02-292D-4897-B20D-869B452A4C0A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{260BBE02-292D-4897-B20D-869B452A4C0A}.Release|Any CPU.Build.0 = Release|Any CPU
-		{773193DB-3981-4577-9108-4D4147C873EB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{773193DB-3981-4577-9108-4D4147C873EB}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{773193DB-3981-4577-9108-4D4147C873EB}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{773193DB-3981-4577-9108-4D4147C873EB}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8F36B825-22A5-4301-AF04-B56D4E2DB4EF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8F36B825-22A5-4301-AF04-B56D4E2DB4EF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8F36B825-22A5-4301-AF04-B56D4E2DB4EF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8F36B825-22A5-4301-AF04-B56D4E2DB4EF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{AF86C3F1-5248-4ED1-98A5-BBDC95A55560}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{AF86C3F1-5248-4ED1-98A5-BBDC95A55560}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{AF86C3F1-5248-4ED1-98A5-BBDC95A55560}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{AF86C3F1-5248-4ED1-98A5-BBDC95A55560}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
